[Silver II] Title: 잃어버린 괄호, Time: 0 ms, Memory: 2024 KB -BaekjoonHub

+++ 백준/Silver/1541. 잃어버린 괄호/잃어버린 괄호.cc
@@ -11,7 +11,7 @@ void Solve()
     string num = "";
     bool minus = false;
 
-    for (int i = 0; i < static_cast<int>(in.size()); i++)
+    for (int i = 0; i < in.size(); i++)
     {
         if (in[i] == '-')
         {
